requirements.txt
# LLM and Agent Framework
# Note: Using versions compatible with agentfence
langchain>=0.3.20,<0.4.0
langchain-core>=0.3.0,<0.4.0
langchain-openai>=0.3.7,<0.4.0
langchain-community>=0.3.19,<0.4.0
langgraph>=0.2.0
litellm>=1.30.0

# Agent Tools
duckduckgo-search>=4.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# Web Framework for A2A endpoints
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0

# Red-teaming Tools
# Note: Install these separately as they may have specific requirements
agentic-radar
agentfence

# Evaluation and Reporting
pandas>=2.1.0
jinja2>=3.1.2
matplotlib>=3.8.0
seaborn>=0.13.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
colorlog>=6.8.0
rich>=13.7.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
